+++ app/api/og-image/route.tsx
@@ -1,6 +1,7 @@
 import { VERCEL_URL } from '@/lib/consts';
 import { ImageResponse } from '@vercel/og';
 import { NextRequest } from 'next/server';
+
 export const runtime = 'edge';
 
 export async function GET(req: NextRequest) {
