+++ lib/utils.ts
@@ -2,6 +2,13 @@ import { type ClassValue, clsx } from 'clsx';
 import { twMerge } from 'tailwind-merge';
 import dayjs from 'dayjs';
 import relativeTime from 'dayjs/plugin/relativeTime';
+import getCastHash from '@/lib/neynar/getCastHash';
+import { stack } from '@/lib/stack/client';
+import { supabaseClient } from '@/lib/supabase/client';
+import findValidEmbed from '@/lib/findValidEmbed';
+import fetchMetadata from '@/lib/fetchMetadata';
+import { CHANNELS } from './consts';
+
 dayjs.extend(relativeTime);
 
 export function cn(...inputs: ClassValue[]) {
@@ -33,4 +40,37 @@ export const isValidNumber = (value: string) => {
 
 export function timeFromNow(date: Date | string) {
   return dayjs(date).fromNow();
+}
+
+export async function getFullHash(username: string, hash: string) {
+  return await getCastHash(`https://warpcast.com/${username}/${hash}`);
+}
+
+export function getHighestRank(validRanks: any[]) {
+  return validRanks.length === 0 ? 0 : Math.min(...validRanks);
+}
+
+export async function getEmbedAndMetadata(fullHash: string) {
+  const { data: cast } = await supabaseClient
+    .from('posts')
+    .select('*')
+    .eq('post_hash', fullHash)
+    .single();
+
+  const embed = findValidEmbed(cast);
+  const url: any = embed?.url;
+  const metadata = await fetchMetadata(url, cast);
+  return { cast, metadata };
+}
+
+export function getChannelData(channelId: any) {
+  return CHANNELS.find((channel) => channel.value === channelId);
+}
+
+export function encodeParams(params: any) {
+  return btoa(JSON.stringify(params));
+}
+
+export function formatPoints(points: any) {
+  return `${(points / 1000).toFixed(1)}K`;
 }
